Mongo form example
This project shows how to take a basic Next.js app and submit data to a
MongoDB backend with a form. It uses `react-form-hook` for the form
state, although `useState` could be used as well.

MongoDB is connected to using localhost in this example, with it running
in a docker container. The collection is built dynamically when first
requested.